SONAR-5931 Cleanup symbol reference API
