+++ sonar-batch/src/main/java/org/sonar/batch/symbol/SymbolData.java
@@ -20,7 +20,7 @@
 
 package org.sonar.batch.symbol;
 
-import org.sonar.api.batch.sensor.symbol.Symbol;
+import org.sonar.api.source.Symbol;
 import org.sonar.batch.index.Data;
 
 import java.util.Collection;
